backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
pandas==2.2.3
numpy==2.2.1
xgboost==2.1.3
scikit-learn==1.6.0
openai==1.58.1
pytest==8.3.4
pytest-asyncio==0.24.0
